#10: add profile page template
