+++ src/pages/components/forumPageWrapper/forumPageWrapper.css
@@ -1,5 +1,6 @@
 .forum-page-wrapper {
   padding: 80px 0;
+  margin: 0 auto;
   display: flex;
   align-items: center;
   justify-content: center;
