+++ src/pages/components/profilePageTemplate/profilePageTemplate.css
@@ -1,4 +1,4 @@
-.profile-page-wrapper {
+.profile-page-template {
   width: 510px;
   margin: 0 auto;
   padding: 60px 0;
