+++ src/pages/components/profilePageTemplate/types.ts
@@ -1,0 +1,3 @@
+export type ProfilePageTemplateProps = {
+  children: React.ReactNode;
+};
