--- src/pages/components/profilePageWrapper/types.ts
@@ -1,3 +1,0 @@
-export type ProfilePageWrapperProps = {
-  children: React.ReactNode;
-};
