+++ src/pages/profile/profile.css
@@ -1,3 +1,3 @@
-.jopa {
+.profile {
   margin: 0;
 }
