+++ src/styles/content.css
@@ -1,5 +1,6 @@
 .content {
   width: 100%;
+  display: flex;
   flex: 1;
 
   &_centered {
